Taint safe objects config: Refactor for consistency, minimal / no effect on performance
